@@ -9,15 +9,15 @@
     "generateLocale": "tsx ./scripts/generateLocale.ts",
     "preview": "nuxt preview",
     "postinstall": "nuxt prepare",
-    "prepare": "husky"
+    "prepare": "husky",
+    "db:generate": "drizzle-kit generate"
   },
   "lint-staged": {
     "*.{ts,js,vue,tsx,jsx}": [
-      "prettier --write",
       "eslint --fix"
     ],
     "*.{json,md}": [
-      "prettier --write"
+      "eslint --fix"
     ]
   },
   "commitlint": {
@@ -26,41 +26,39 @@
     ]
   },
   "devDependencies": {
-    "@commitlint/cli": "^19.2.1",
-    "@commitlint/config-conventional": "^19.1.0",
-    "@iconify/json": "^2.2.196",
-    "@iconify/tailwind": "^0.1.4",
-    "@nuxt/devtools": "^1.1.5",
-    "@nuxtjs/color-mode": "^3.3.3",
-    "@nuxtjs/eslint-module": "^4.1.0",
+    "@commitlint/cli": "^19.5.0",
+    "@commitlint/config-conventional": "^19.5.0",
+    "@iconify/json": "^2.2.253",
+    "@iconify/tailwind": "^1.1.3",
+    "@nuxt/devtools": "^1.5.1",
+    "@nuxt/eslint": "^0.5.7",
+    "@nuxtjs/color-mode": "^3.5.1",
     "@nuxtjs/google-fonts": "^3.2.0",
-    "@nuxtjs/i18n": "^8.2.0",
-    "@nuxtjs/seo": "^2.0.0-rc.10",
-    "@nuxtjs/tailwindcss": "^6.11.4",
-    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@nuxtjs/i18n": "^8.5.5",
+    "@nuxtjs/seo": "^2.0.0-rc.21",
+    "@nuxtjs/tailwindcss": "^6.12.1",
     "dotenv": "^16.4.5",
-    "eslint": "^8.57.0",
-    "eslint-config-love": "^43.1.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-vue": "^9.24.0",
-    "eventsource-parser": "^1.1.2",
-    "husky": "^9.0.11",
+    "eslint": "^9.11.1",
+    "eventsource-parser": "^2.0.1",
+    "husky": "^9.1.6",
     "json-stringify-pretty-compact": "^4.0.0",
-    "nuxt": "^3.11.1",
-    "nuxt-gtag": "^2.0.5",
-    "prettier": "^3.2.5",
-    "tsx": "^4.7.1",
-    "typescript": "^5.4.3",
-    "undici": "^6.10.2",
-    "vue": "^3.4.21",
-    "vue-eslint-parser": "^9.4.2",
-    "vue-router": "^4.3.0",
-    "zx": "^7.2.3"
+    "nuxt": "^3.13.2",
+    "nuxt-gtag": "^3.0.1",
+    "tsx": "^4.19.1",
+    "typescript": "^5.6.2",
+    "undici": "^6.19.8",
+    "vue": "^3.5.9",
+    "vue-router": "^4.4.5",
+    "wrangler": "^3.75.0",
+    "zx": "^8.1.8"
   },
   "dependencies": {
-    "@vueuse/core": "^10.9.0"
+    "@vueuse/core": "^11.1.0",
+    "@nuxthub/core": "^0.7.10",
+    "drizzle-kit": "^0.24.2",
+    "drizzle-orm": "0.33.0",
+    "h3-zod": "^0.5.3",
+    "nuxt-auth-utils": "^0.3.7",
+    "zod": "^3.23.8"
   }
 }
