@@ -1,10 +1,10 @@
-import en from '../lang/en.json'
-import zh from '../lang/zh-hans.json'
-import { localePrompt } from './prompts'
 import { promises as fs } from 'node:fs'
-import { chatGPT } from './openAI'
 import stringify from 'json-stringify-pretty-compact'
-import locale from '../locale'
+import en from '../locales/en.json'
+import zh from '../locales/zh-hans.json'
+import index from '../locales'
+import { localePrompt } from './prompts'
+import { chatGPT } from './openAI'
 
 const getDiffKeys = (obj1: any, obj2: any) => {
   return Object.keys(obj2).reduce((diff: any, key) => {
@@ -15,13 +15,14 @@ const getDiffKeys = (obj1: any, obj2: any) => {
     return diff
   }, {})
 }
-async function main (lang: string) {
-  const file = `./lang/${lang}.json`
+async function main(lang: string) {
+  const file = `./locales/${lang}.json`
   let exists: Record<string, any> = {}
   try {
     const existsStr = await fs.readFile(file, { encoding: 'utf-8' })
     exists = JSON.parse(existsStr)
-  } catch (e: any) {
+  }
+  catch (e: any) {
     console.log(`read error：${e.message}`)
   }
   const enDiff = getDiffKeys(exists, en)
@@ -38,12 +39,12 @@ async function main (lang: string) {
     exists[key] = data[key]
   }
 
-  await fs.writeFile(`./lang/${lang}.json`, stringify(exists))
+  await fs.writeFile(`./locales/${lang}.json`, stringify(exists))
 }
 
 // const badResult = ['da', 'fi', 'he', 'hu', 'id', 'it', 'th', 'tr'].map(lang => ({ code: lang }))
-async function go () {
-  for (const l of locale) {
+async function go() {
+  for (const l of index) {
     if (l.code !== 'en' && l.code !== 'zh-hans') {
       await main(l.code)
     }
