@@ -1,4 +1,4 @@
-export default defineEventHandler(ev => {
+export default defineEventHandler((ev) => {
   const query = getQuery(ev)
   const locale = query.locale as string
   console.log(locale)
